refactor(icons): keep flat folder structure for /images since imagemin-cli does not preserve nested folder structures
