+++ lib/components/Select/Select.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
-import IconArrowDown from './../../images/icons/arrow_down.svg';
+import IconArrowDown from './../../images/icon_arrow_down.svg';
 
 require(`./Select.${process.env.NODE_ENV === 'storybook' ? 'scss' : 'css'}`);
