+++ lib/components/UserStatus/UserStatus.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
 import AuthNav from '../AuthNav/AuthNav';
-import IconUser from './../../images/icons/user.svg';
+import IconUser from './../../images/icon_user.svg';
 
 require(`./UserStatus.${process.env.NODE_ENV === 'storybook' ? 'scss' : 'css'}`);
